requirements.txt
discord.py
python-dotenv
watchdog
google-generativeai
google-genai
